Move mac-specific classes to a subfolder
