+++ RepoZ.UI.Mac.Story/NativeSupport/MacThreadDispatcher.cs
@@ -3,7 +3,7 @@ using AppKit;
 using Foundation;
 using RepoZ.Api.Common;
 
-namespace RepoZ.UI.Mac.Story
+namespace RepoZ.UI.Mac.Story.NativeSupport
 {
     public class MacThreadDispatcher : IThreadDispatcher
     {
